refactor: mv shared/assets/icons
