+++ src/shared/assets/icons/CheckmarkOutlineIcon.tsx
