+++ src/shared/assets/icons/index.ts
